server/requirements.txt
alembic==1.7.3
aniso8601==9.0.1
appnope==0.1.2
attrs==21.2.0
backcall==0.2.0
black==21.9b0
blis==0.7.4
boto3==1.18.56
botocore==1.21.56
catalogue==2.0.6
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
cymem==2.0.5
debugpy==1.5.0
decorator==5.1.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0-py3-none-any.whl
entrypoints==0.3
Flask==2.0.1
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.2
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.1.0
jsonschema==4.0.1
jupyter-client==7.0.6
jupyter-core==4.8.1
Mako==1.1.5
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
murmurhash==1.0.5
mypy-extensions==0.4.3
mysql-connector-python==8.0.26
nest-asyncio==1.5.1
nltk==3.6.5
numpy==1.21.2
packaging==21.0
pandas==1.3.3
parso==0.8.2
pathspec==0.9.0
pathy==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
preshed==3.0.5
prompt-toolkit==3.0.20
protobuf==3.18.0
ptyprocess==0.7.0
pydantic==1.8.2
Pygments==2.10.0
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.3
pyzmq==22.3.0
regex==2021.9.30
requests==2.26.0
s3transfer==0.5.0
selenium==3.141.0
six==1.16.0
smart-open==5.2.1
spacey==0.1.1
spacy==3.1.3
spacy-legacy==3.0.8
SQLAlchemy==1.4.25
srsly==2.4.1
thinc==8.0.10
tomli==1.2.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wasabi==0.8.2
wcwidth==0.2.5
Werkzeug==2.0.1
